Build buy realty token module
